backend/pyproject.toml
[project]
name = "gazabot-backend"
version = "0.1.0"
description = "Guardian dashboard backend for Gazabot Browser Use integration."
requires-python = ">=3.11"
dependencies = [
  "browser-use-sdk",
  "croniter>=3.0,<4.0",
  "fastapi>=0.115,<1.0",
  "httpx>=0.27,<1.0",
  "pydantic-settings>=2.6,<3.0",
  "sqlmodel>=0.0.22,<0.0.25",
  "uvicorn>=0.30,<1.0",
]

[dependency-groups]
dev = [
  "pytest>=8.3,<9.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
